+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/.gitignore
