+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/.openapi-generator/FILES
