+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/.travis.yml
