+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/api/openapi.yaml
