+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/cluster-api.yaml
