+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/components.yaml
