+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/BrokerState.md
