+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/Error.md
