+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/Operation.md
