+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/PartitionConfig.md
