+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/PartitionState.md
