+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/PartitionStateCode.md
