+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/docs/PlannedOperationsResponse.md
