+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/git_push.sh
