+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/model_broker_state.go
