+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/model_error.go
