+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/model_partition_config.go
