+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/model_planned_operations_response.go
