+++ camunda-scaling-operator/pkg/scalingclient/zbmgmt/model_topology_change.go
